requirements.txt
python-dotenv~=1.0.1
requests~=2.32.3
google-generativeai~=0.6
google~=3.0.0
retry~=0.9.2
protobuf
proto-plus==1.24.0.dev1
grpcio-testing
grpcio
grpcio-tools